# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty 100 Financial Intelligence API",
    "description": "Production REST API providing endpoints for stock screening, peer analytics, company profiles, financial statements, clustering archetypes, and documents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Companies",
        "description": "Returns list of companies with roe_pct, roce_pct, and sector metadata.",
        "operationId": "get_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by broad sector",
              "title": "Sector"
            },
            "description": "Filter by broad sector"
          },
          {
            "name": "market_cap_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by market cap category",
              "title": "Market Cap Category"
            },
            "description": "Filter by market cap category"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by partial name or ticker symbol",
              "title": "Search"
            },
            "description": "Search by partial name or ticker symbol"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Company Profile",
        "description": "Returns full company profile including company details, sector data, and latest year KPIs.",
        "operationId": "get_company_profile_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/pl": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Profit And Loss",
        "description": "Returns Profit & Loss statement history array for the given company.",
        "operationId": "get_profit_and_loss_api_v1_companies__ticker__pl_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start year filter",
              "title": "From Year"
            },
            "description": "Start year filter"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End year filter",
              "title": "To Year"
            },
            "description": "End year filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/bs": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Balance Sheet",
        "description": "Returns Balance Sheet statement history array for the given company.",
        "operationId": "get_balance_sheet_api_v1_companies__ticker__bs_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start year filter",
              "title": "From Year"
            },
            "description": "Start year filter"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End year filter",
              "title": "To Year"
            },
            "description": "End year filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/cashflow": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Cashflow",
        "description": "Returns Cash Flow statement history array for the given company.",
        "operationId": "get_cashflow_api_v1_companies__ticker__cashflow_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start year filter",
              "title": "From Year"
            },
            "description": "Start year filter"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End year filter",
              "title": "To Year"
            },
            "description": "End year filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/ratios": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Financial Ratios",
        "description": "Returns all computed financial KPIs per year for the target company.",
        "operationId": "get_financial_ratios_api_v1_companies__ticker__ratios_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Single year filter",
              "title": "Year"
            },
            "description": "Single year filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Tearsheet Pdf",
        "description": "Downloads the pre-generated executive 2-page PDF tearsheet for the target company.",
        "operationId": "get_tearsheet_pdf_api_v1_companies__ticker__tearsheet_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Company Documents",
        "description": "Returns annual report links with is_url_valid boolean flag for the given company.",
        "operationId": "get_company_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}